feature flag. fixed application failed to load error
