+++ src/main/java/uk/gov/companieshouse/pscdataapi/repository/CompanyPscRepository.java
@@ -7,7 +7,6 @@ import java.util.Optional;
 import org.springframework.data.mongodb.repository.Aggregation;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
-import uk.gov.companieshouse.api.model.PscStatementDocument;
 import uk.gov.companieshouse.pscdataapi.models.PscDocument;
 
 public interface CompanyPscRepository extends MongoRepository<PscDocument, String> {
