+++ src/test/resources/application.properties
@@ -9,3 +9,14 @@ api.api-key=${CHS_API_KEY:chsApiKey}
 feature.seeding_collection_enabled=false
 
 items-per-page-max-internal=500
+
+management.endpoints.web.base-path=/psc-data-api
+management.endpoints.web.path-mapping.health=/healthcheck
+chs.api.kafka.url=${CHS_KAFKA_API_URL:localhost}
+chs.api.kafka.resource-changed.uri=${PSC_API_RESOURCE_CHANGED_URI:/private/resource-changed}
+chs.api.metrics.url=${API_LOCAL_URL:localhost}
+logger.namespace=psc-data-api
+spring.data.mongodb.uri=${MONGODB_URL:mongodb://mongo:27017}
+spring.data.mongodb.name=company_pscs
+spring.jackson.default-property-inclusion=NON_NULL
+mongodb.pscs.collection.name=delta_company_pscs
