--- src/test/resources/application.yml
@@ -1,31 +1,0 @@
-management:
-  endpoints:
-    web:
-      base-path: /psc-data-api
-      path-mapping:
-        health: /healthcheck
-
-chs:
-  api:
-    kafka:
-      url: ${CHS_KAFKA_API_URL:localhost}
-      resource-changed:
-        uri: ${PSC_API_RESOURCE_CHANGED_URI:/private/resource-changed}
-    metrics:
-      url: ${API_LOCAL_URL:localhost}
-
-logger:
-  namespace: psc-data-api
-
-spring:
-  data:
-    mongodb:
-      uri: ${MONGODB_URL:mongodb://mongo:27017}
-      name: company_pscs
-  jackson:
-    default-property-inclusion: NON_NULL
-
-mongodb:
-  pscs:
-    collection:
-      name: delta_company_pscs
